Fix full fluid transmitters getting stuck full
If a fluid transmitter ends up filling up completely it'll have an
integer amount of fluid in it, which resulted in the input code skipping
it as there would be no fluid left after taking the floor of the amount
from the tank.

Fix by taking the ceiling of the stored amount minus one.  Fixes #40.