Fix crash on deconstruct from editor
When using the deconstruction planner from the editor the game raises
the script_raised_destroy event for the entities removed.  This was not
handled and in combination with the recepie being read before
entity.valid was checked caused a non-recoverable error when fluid
extractors where removed with a deconstruction planner in the editor.

Fix by handling script_raised_destroy and move the entity.valid check
before trying to read the recipe of the fluid extractor.