Add support for building multiple versions
Add support for a "variants" array in the info.json file which specifies
a list of mod variants to build along with variant specific overrides in
the info.json file.  Add support for an "additional_files" mapping in
the info.json file for specifying extra files that will be written to
the mod.  Combined with the variants option this allows making the mod
for multiple Factorio versions with code specific to each Factorio
version.